requirements.txt
flask
flask-mysqldb
python-dotenv
pyjwt
bcrypt
flask-cors